+++ pandas/tests/frame/indexing/test_getitem.py
@@ -1,6 +1,8 @@
+import numpy as np
 import pytest
 
-from pandas import DataFrame, MultiIndex
+from pandas import DataFrame, MultiIndex, period_range
+import pandas._testing as tm
 
 
 class TestGetitem:
@@ -14,3 +16,16 @@ class TestGetitem:
 
         with pytest.raises(KeyError, match="notevenone"):
             df["notevenone"]
+
+    def test_getitem_periodindex(self):
+        rng = period_range("1/1/2000", periods=5)
+        df = DataFrame(np.random.randn(10, 5), columns=rng)
+
+        ts = df[rng[0]]
+        tm.assert_series_equal(ts, df.iloc[:, 0])
+
+        # GH#1211; smoketest unrelated to the rest of this test
+        repr(df)
+
+        ts = df["1/1/2000"]
+        tm.assert_series_equal(ts, df.iloc[:, 0])
