+++ pandas/tests/series/test_timeseries.py
@@ -1,5 +1,4 @@
 import numpy as np
-import pytest
 
 import pandas as pd
 from pandas import DataFrame, Series, date_range, timedelta_range
@@ -70,37 +69,3 @@ class TestTimeSeries:
             ]
         )
         tm.assert_series_equal(result, expected)
-
-    @pytest.mark.parametrize("tz", [None, "US/Central"])
-    def test_asarray_object_dt64(self, tz):
-        ser = Series(pd.date_range("2000", periods=2, tz=tz))
-
-        with tm.assert_produces_warning(None):
-            # Future behavior (for tzaware case) with no warning
-            result = np.asarray(ser, dtype=object)
-
-        expected = np.array(
-            [pd.Timestamp("2000-01-01", tz=tz), pd.Timestamp("2000-01-02", tz=tz)]
-        )
-        tm.assert_numpy_array_equal(result, expected)
-
-    def test_asarray_tz_naive(self):
-        # This shouldn't produce a warning.
-        ser = Series(pd.date_range("2000", periods=2))
-        expected = np.array(["2000-01-01", "2000-01-02"], dtype="M8[ns]")
-        result = np.asarray(ser)
-
-        tm.assert_numpy_array_equal(result, expected)
-
-    def test_asarray_tz_aware(self):
-        tz = "US/Central"
-        ser = Series(pd.date_range("2000", periods=2, tz=tz))
-        expected = np.array(["2000-01-01T06", "2000-01-02T06"], dtype="M8[ns]")
-        result = np.asarray(ser, dtype="datetime64[ns]")
-
-        tm.assert_numpy_array_equal(result, expected)
-
-        # Old behavior with no warning
-        result = np.asarray(ser, dtype="M8[ns]")
-
-        tm.assert_numpy_array_equal(result, expected)
